Added option to treat missing departure timestamps as censored data.
By default, missing departure timestamps are assumed to mean entity is
still in system. For occupancy computations, the entity is treated as
departing after the end of the analysis range. In this way they
count for occupancy but do not count for departures within the
analysis range.

If you want records with missing departure timestamps to simply
be ignored (i.e. assumed to be bad data), just set the --no_censored_departures
flag or use no_censored_departures=True when calling make_hills().